fix: share iam_policy_document() validation via carina-aws-types
Move IAM policy document structural validation (version, statement,
effect, principal, condition) from carina-provider-awscc into
carina-aws-types so both AWS and AWSCC providers use the same
validation logic instead of the AWS provider's unvalidated Map(String).

Closes #557
